creator-brief: lighten the skill to a short SKILL.md plus references
SKILL.md drops from 5,343 to 1,908 words. It keeps only what a run needs
and points at transcript-mining.md, profile-spec.md and evidence-rules.md
for the detail, instead of restating them. History, timings and war
stories are gone; the plugin's skill root is stated so commands work from
any directory.

Files: 24 -> 19.
- classify_gems.py (the scripted extractor) and compare_extracts.py (a
  one-off benchmark) removed with their tests; nothing imported them.
- corpus_io.py + ledger_io.py merged into store_io.py; quote_timestamp's
  locate() folded into verify_quotes.py; sponsor_spans folded into
  fetch_cues.py.
- references/gem-classifier.md renamed extractor-rubric.md; the agent
  file keeps its registered name.

Fixes:
- channel_context.py --write-context writes the extractor context block
  from the saved full context plus the model's format call, replacing
  the untyped python -c placeholder in step 2; label is an enum.
- merge pass described one way everywhere: sharded by default.
- batch size is derived from the agent cap, no file says 25 any more.
- concurrency guidance back to the standard 20 agents; 40 was measured
  slower.
- Mode B -> CONNECT in every reference; stale ledger-view mentions and
  the build_html.py page description brought up to the current page.

779 tests pass.
