+++ tests/test_creator_brief_scripts.py
@@ -1,23 +1,17 @@
-"""The creator-brief scripts: local verification, the deterministic
-connections page, and the scripted extractor's mechanical guarantees (return
-files, resume, retry, key resolution, loud no-key exit). The retrieval and
-assembly stages have their own files — ``test_fetch_cues.py`` and
-``test_assemble_extracts.py``. No real network anywhere: the scripted
-extractor talks to a fake OpenAI-compatible endpoint on 127.0.0.1.
+"""The creator-brief scripts: local quote verification and the deterministic
+connections page. The retrieval and assembly stages have their own files,
+``test_fetch_cues.py`` and ``test_assemble_extracts.py``. No real network.
 """
 
-import http.server
 import json
 import subprocess
 import sys
-import threading
 from pathlib import Path
 
 _SCRIPTS = (Path(__file__).resolve().parents[1]
             / "skills" / "tl-creator-brief" / "scripts")
 sys.path.insert(0, str(_SCRIPTS))
-import classify_gems  # noqa: E402
-import ledger_io  # noqa: E402
+import store_io  # noqa: E402
 
 
 def _write_jsonl(path: Path, rows: list[dict]) -> Path:
@@ -98,7 +92,7 @@ def test_a_meta_header_is_not_a_candidate_and_survives_the_pass(tmp_path):
     header = {"schema": "tl-creator-meta/v2", "channel_id": 1, "channel_name": "P",
               "coverage": {"facts": 1}}
     ledger = tmp_path / "1-facts.jsonl"
-    ledger_io.write_ledger(ledger, header, [
+    store_io.write_ledger(ledger, header, [
         {"provenance": "transcript", "video": "1:vid1",
          "quote": "I grew up in a tiny town in Ohio"}])
     proc = subprocess.run(
@@ -108,7 +102,7 @@ def test_a_meta_header_is_not_a_candidate_and_survives_the_pass(tmp_path):
         capture_output=True, text=True)
     assert proc.returncode == 0
     assert json.loads(proc.stdout)["candidates"] == 1      # the header is not one
-    meta, facts = ledger_io.read_ledger(tmp_path / "out.jsonl")
+    meta, facts = store_io.read_ledger(tmp_path / "out.jsonl")
     assert meta == header
     assert len(facts) == 1 and facts[0]["verify"]["match"] == "exact"
 
@@ -164,7 +158,7 @@ _CONN_MD = (
 
 def _write_ledger(tmp_path: Path, facts=None, meta=None) -> Path:
     path = tmp_path / "42-facts.jsonl"
-    ledger_io.write_ledger(path, _META if meta is None else meta,
+    store_io.write_ledger(path, _META if meta is None else meta,
                            _FACTS if facts is None else facts)
     return path
 
@@ -282,7 +276,7 @@ def test_a_legacy_headerless_ledger_still_renders_with_meta(tmp_path):
     src = tmp_path / "42-7-connections.md"
     src.write_text(_CONN_MD)
     facts = tmp_path / "42-facts.jsonl"
-    ledger_io.write_ledger(facts, None, _FACTS)
+    store_io.write_ledger(facts, None, _FACTS)
     meta = tmp_path / "42-meta.json"
     meta.write_text(json.dumps(_META))
     subprocess.run([sys.executable, str(_SCRIPTS / "build_html.py"), "--in", str(src),
@@ -361,283 +355,9 @@ def test_who_they_are_link_only_http_schemes(tmp_path):
     assert 'href="javascript' not in who and "grew up in Ohio" in who
 
 
-# --------------------------------------------------------------------------- #
-# classify_gems.py — the scripted extractor, against a local fake endpoint
-# --------------------------------------------------------------------------- #
-def _windows_in(prompt: str) -> list[dict]:
-    """The window array the rendered extractor message carries."""
-    tail = prompt.split("=== WINDOWS", 1)[1].split("===\n", 1)[1]
-    return json.loads(tail.splitlines()[0])
-
-
-def _extract_for(prompt: str, **over) -> dict:
-    """A contract-shaped extract for the batch in this prompt: every other
-    window a gem, each verdict echoing its own window's `start`."""
-    gems, not_gems = [], []
-    for w in _windows_in(prompt):
-        if w["i"] % 2 == 0:
-            text = w["text"]
-            ws = text.split()
-            gems.append({"i": w["i"], "start": w["start"],
-                         "anchor": " ".join(ws[:5]),
-                         "life_domain": "family", "speaker_guess": "host",
-                         "sensitivity": "none", "entity_corrections": {},
-                         "notable": "father ran a bakery",
-                         "claim": "father ran a bakery",
-                         "quote_span": {"first": " ".join(ws[:4]),
-                                        "last": " ".join(ws[-4:])},
-                         "confidence": "confirmed"})
-        else:
-            not_gems.append({"i": w["i"], "speaker_guess": "guest",
-                             "reason": "third-party"})
-    out = {"gems": gems, "not_gems": not_gems}
-    out.update(over)
-    return out
-
-
-class _FakeLLM:
-    """An OpenAI-compatible /chat/completions endpoint on 127.0.0.1."""
-
-    def __init__(self, content):
-        self.content = content       # callable(prompt) -> response content
-        self.requests = []
-        handler = self._handler()
-        self.httpd = http.server.HTTPServer(("127.0.0.1", 0), handler)
-        self.thread = threading.Thread(target=self.httpd.serve_forever,
-                                       daemon=True)
-        self.thread.start()
-
-    @property
-    def url(self) -> str:
-        return f"http://127.0.0.1:{self.httpd.server_address[1]}/v1"
-
-    def _handler(server_self):
-        class H(http.server.BaseHTTPRequestHandler):
-            def log_message(self, *a):
-                pass
-
-            def do_POST(self):
-                body = json.loads(self.rfile.read(
-                    int(self.headers["Content-Length"])).decode())
-                prompt = body["messages"][0]["content"]
-                server_self.requests.append(body)
-                content = server_self.content(prompt)
-                payload = json.dumps({
-                    "choices": [{"message": {"content": content}}],
-                    "usage": {"prompt_tokens": 11, "completion_tokens": 3},
-                }).encode()
-                self.send_response(200)
-                self.send_header("Content-Type", "application/json")
-                self.send_header("Content-Length", str(len(payload)))
-                self.end_headers()
-                self.wfile.write(payload)
-        return H
-
-    def stop(self):
-        self.httpd.shutdown()
-        self.httpd.server_close()
-
-
-def _corpus(tmp_path: Path, batches: int = 2, per: int = 2):
-    root = tmp_path / "batches"
-    root.mkdir()
-    for n in range(1, batches + 1):
-        root.joinpath(f"batch-{n:03d}.json").write_text(json.dumps([
-            {"id": f"1:vid{n}", "start": 10 * n + i,
-             "text": f"my dad ran a bakery in a tiny town number {n}{i}",
-             "title": "t", "format_hint": None, "in_sponsor_read": False}
-            for i in range(per)]))
-    ctx = tmp_path / "context.json"
-    ctx.write_text(json.dumps({"channel_name": "Patterrz"}))
-    return root, ctx
-
-
-def _run_extract(batch_dir, ctx, *extra, env=None):
-    e = {"PATH": "/usr/bin:/bin"}
-    e.update(env or {})
-    return subprocess.run(
-        [sys.executable, str(_SCRIPTS / "classify_gems.py"),
-         "--batches", str(batch_dir), "--context", str(ctx), *extra],
-        capture_output=True, text=True, env=e)
-
-
-def test_scripted_extractor_writes_one_return_file_per_batch(tmp_path):
-    batches, ctx = _corpus(tmp_path)
-    fake = _FakeLLM(lambda p: json.dumps(_extract_for(p)))
-    try:
-        proc = _run_extract(batches, ctx, env={
-            "CREATOR_BRIEF_LLM_API_KEY": "k",
-            "CREATOR_BRIEF_LLM_BASE_URL": fake.url,
-            "CREATOR_BRIEF_LLM_MODEL": "test-model"})
-    finally:
-        fake.stop()
-    assert proc.returncode == 0, proc.stderr
-    returns = tmp_path / "returns"
-    files = sorted(p.name for p in returns.glob("*.json"))
-    assert files == ["batch-001.extract.json", "batch-002.extract.json"]
-    obj = json.loads((returns / "batch-002.extract.json").read_text())
-    # the response carried neither key; both are filled in from the batch
-    assert obj["batch"] == "002" and obj["windows"] == 2
-    assert [g["start"] for g in obj["gems"]] == [20]     # echoed, assemblable
-    assert [x["i"] for x in obj["not_gems"]] == [1]
-    summary = json.loads(proc.stdout)
-    assert summary["batches"] == 2 and summary["batches_written"] == 2
-    assert summary["errors"] == 0 and summary["windows"] == 4
-    assert summary["skipped_existing"] == 0
-    assert summary["prompt_tokens"] == 22 and summary["completion_tokens"] == 6
-    assert summary["largest_return_chars"] > 100
-    line = next(ln for ln in proc.stderr.splitlines()
-                if ln.startswith("FUNNEL stage=extract"))
-    f = dict(kv.split("=", 1) for kv in line.split()[1:])
-    assert f["path"] == "api" and f["batches"] == "2" and f["windows"] == "4"
-    assert f["written"] == "2" and f["errors"] == "0"
-
-
-def test_existing_return_files_are_skipped_unless_forced(tmp_path):
-    batches, ctx = _corpus(tmp_path, batches=2)
-    returns = tmp_path / "returns"
-    returns.mkdir()
-    (returns / "batch-001.extract.json").write_text('{"gems": [], '
-                                                    '"not_gems": []}')
-    fake = _FakeLLM(lambda p: json.dumps(_extract_for(p)))
-    env = {"CREATOR_BRIEF_LLM_API_KEY": "k",
-           "CREATOR_BRIEF_LLM_BASE_URL": fake.url,
-           "CREATOR_BRIEF_LLM_MODEL": "test-model"}
-    try:
-        first = _run_extract(batches, ctx, env=env)
-        assert json.loads(first.stdout)["skipped_existing"] == 1
-        assert json.loads(first.stdout)["batches_written"] == 1
-        assert len(fake.requests) == 1
-        forced = _run_extract(batches, ctx, "--force", env=env)
-    finally:
-        fake.stop()
-    assert json.loads(forced.stdout)["skipped_existing"] == 0
-    assert json.loads(forced.stdout)["batches_written"] == 2
-    assert len(fake.requests) == 3
-    # --force actually replaced the stub
-    assert json.loads(
-        (returns / "batch-001.extract.json").read_text())["gems"]
-
-
-def test_malformed_content_is_retried_then_counted_as_an_error(tmp_path):
-    batches, ctx = _corpus(tmp_path, batches=1)
-    fake = _FakeLLM(lambda p: json.dumps(["not", "an", "object"]))
-    try:
-        proc = _run_extract(batches, ctx, env={
-            "CREATOR_BRIEF_LLM_API_KEY": "k",
-            "CREATOR_BRIEF_LLM_BASE_URL": fake.url,
-            "CREATOR_BRIEF_LLM_MODEL": "test-model"})
-        attempts = len(fake.requests)
-    finally:
-        fake.stop()
-    assert attempts == 1 + classify_gems.RETRIES == 3
-    assert proc.returncode == 1
-    summary = json.loads(proc.stdout)
-    assert summary["errors"] == 1 and summary["batches_written"] == 0
-    # nothing written: the assembler sees a missing return file for the batch
-    assert not list((tmp_path / "returns").glob("*.json"))
-
-
-def test_a_fenced_object_is_accepted(tmp_path):
-    batches, ctx = _corpus(tmp_path, batches=1)
-    fake = _FakeLLM(
-        lambda p: "```json\n" + json.dumps(_extract_for(p)) + "\n```")
-    try:
-        proc = _run_extract(batches, ctx, env={
-            "CREATOR_BRIEF_LLM_API_KEY": "k",
-            "CREATOR_BRIEF_LLM_BASE_URL": fake.url,
-            "CREATOR_BRIEF_LLM_MODEL": "test-model"})
-    finally:
-        fake.stop()
-    assert proc.returncode == 0, proc.stderr
-    obj = json.loads(
-        (tmp_path / "returns" / "batch-001.extract.json").read_text())
-    assert obj["batch"] == "001" and len(obj["gems"]) == 1
-
-
-def _run_no_key(tmp_path: Path, windows: list[dict], env: dict | None = None):
-    batches = tmp_path / "batches"
-    batches.mkdir()
-    (batches / "batch-000.json").write_text(json.dumps(windows))
-    ctx = tmp_path / "context.json"
-    ctx.write_text("{}")
-    e = {"PATH": "/usr/bin:/bin"}
-    e.update(env or {})
-    return subprocess.run(
-        [sys.executable, str(_SCRIPTS / "classify_gems.py"),
-         "--batches", str(batches), "--context", str(ctx)],
-        capture_output=True, text=True,
-        env=e), batches
-
-
-def test_no_api_key_emits_the_fallback_marker_and_its_own_exit_code(tmp_path):
-    proc, batches = _run_no_key(tmp_path, [{"id": "1:aaa", "start": 5,
-                                            "text": "my dad ran a bakery"}])
-    # a code of its own: not argparse's usage 2, not the "errors" 1
-    assert proc.returncode == classify_gems.EXIT_FALLBACK_REQUIRED == 20
-    marker = next(ln for ln in proc.stderr.splitlines()
-                  if ln.startswith("FALLBACK_REQUIRED"))
-    fields = dict(kv.split("=", 1) for kv in marker.split()[1:])
-    assert fields["reason"] == "missing_api_key"
-    # the fallback consumes exactly these files, so the path must resolve
-    assert Path(fields["batches_dir"]) == batches.resolve()
-    assert fields["batch_files"] == "1" and fields["windows"] == "1"
-    assert "CREATOR_BRIEF_LLM_API_KEY" in proc.stderr
-    assert "CREATOR_BRIEF_LLM_BASE_URL" in proc.stderr
-    assert "CREATOR_BRIEF_LLM_MODEL" in proc.stderr
-    assert "fall back" in proc.stderr
-
-
-def test_missing_model_alone_also_exits_fallback_required(tmp_path):
-    proc, _ = _run_no_key(
-        tmp_path, [{"id": "1:aaa", "start": 5, "text": "my dad ran a bakery"}],
-        env={"CREATOR_BRIEF_LLM_API_KEY": "k",
-             "CREATOR_BRIEF_LLM_BASE_URL": "http://127.0.0.1:1"})
-    assert proc.returncode == classify_gems.EXIT_FALLBACK_REQUIRED == 20
-    marker = next(ln for ln in proc.stderr.splitlines()
-                  if ln.startswith("FALLBACK_REQUIRED"))
-    fields = dict(kv.split("=", 1) for kv in marker.split()[1:])
-    assert fields["missing"] == "CREATOR_BRIEF_LLM_MODEL"
-    assert "CREATOR_BRIEF_LLM_MODEL" in proc.stderr
-
-
-def test_no_api_key_still_prints_its_funnel_line(tmp_path):
-    proc, _ = _run_no_key(tmp_path, [{"id": "1:aaa", "start": 5, "text": "x"},
-                                     {"id": "1:bbb", "start": 9, "text": "y"}])
-    line = next(ln for ln in proc.stderr.splitlines()
-                if ln.startswith("FUNNEL stage=extract"))
-    fields = dict(kv.split("=", 1) for kv in line.split()[1:])
-    assert fields["path"] == "fallback_required"
-    assert fields["windows"] == "2" and fields["written"] == "0"
-    assert float(fields["elapsed_s"]) >= 0
-
-
-def test_concurrency_comes_from_the_env_within_bounds(monkeypatch):
-    monkeypatch.delenv("CREATOR_BRIEF_LLM_CONCURRENCY", raising=False)
-    assert classify_gems.env_concurrency() == classify_gems.CONCURRENCY == 16
-    monkeypatch.setenv("CREATOR_BRIEF_LLM_CONCURRENCY", "8")
-    assert classify_gems.env_concurrency() == 8
-    monkeypatch.setenv("CREATOR_BRIEF_LLM_CONCURRENCY", "9999")
-    assert classify_gems.env_concurrency() == classify_gems.MAX_CONCURRENCY
-    monkeypatch.setenv("CREATOR_BRIEF_LLM_CONCURRENCY", "0")
-    assert classify_gems.env_concurrency() == classify_gems.MIN_CONCURRENCY
-    monkeypatch.setenv("CREATOR_BRIEF_LLM_CONCURRENCY", "lots")
-    assert classify_gems.env_concurrency() == classify_gems.CONCURRENCY
-
-
-def test_the_extract_parser_fills_in_batch_and_windows_only_when_missing():
-    got = classify_gems.parse_extract('{"gems": [], "not_gems": []}', "007", 25)
-    assert got == {"batch": "007", "windows": 25, "gems": [], "not_gems": []}
-    kept = classify_gems.parse_extract(
-        '{"batch": "003", "windows": 4, "gems": [], "not_gems": []}', "007", 25)
-    assert kept["batch"] == "003" and kept["windows"] == 4
-    for bad in ('[]', 'not json', '{"gems": []}', '{"gems": {}, "not_gems": []}'):
-        assert classify_gems.parse_extract(bad, "007", 25) is None
-
-
 def test_locate_prefers_the_occurrence_nearest_the_hint():
     sys.path.insert(0, str(_SCRIPTS))
-    from quote_timestamp import locate
+    from verify_quotes import locate
     cues = [(10.0, "I grew up in Ohio you know"),
             (200.0, "and then she said I grew up in Ohio too")]
     quote = "I grew up in Ohio"
@@ -656,6 +376,29 @@ def test_rows_raises_on_withheld_premium_fields():
                        "_upgrade_required": {"message": "upgrade",
                                              "fields": ["transcript"]}})
     assert tl_data._rows({"results": [{"id": 1}]}) == [{"id": 1}]
+
+
+def test_write_context_builds_the_extractor_block_from_the_saved_full_context(tmp_path):
+    full = tmp_path / "context-full.json"
+    full.write_text(json.dumps({"name": "Ali Abdaal",
+                                "context_stats": {"fp_per_1k_words_median": 41}}))
+    out = tmp_path / "context.json"
+    proc = subprocess.run(
+        [sys.executable, str(_SCRIPTS / "channel_context.py"), "--from", str(full),
+         "--format-label", "solo", "--format-evidence", "fp density 41/1k",
+         "--known-facts", "ex-doctor; lives in London", "--write-context", str(out)],
+        capture_output=True, text=True)
+    assert proc.returncode == 0, proc.stderr
+    assert json.loads(out.read_text()) == {
+        "channel_name": "Ali Abdaal", "host_names": ["Ali Abdaal"],
+        "known_facts": ["ex-doctor", "lives in London"],
+        "format_label": "solo", "format_evidence": "fp density 41/1k"}
+    # the label is an enum: a near-miss fails here, not in front of 20 agents
+    bad = subprocess.run(
+        [sys.executable, str(_SCRIPTS / "channel_context.py"), "--from", str(full),
+         "--format-label", "Solo", "--write-context", str(out)],
+        capture_output=True, text=True)
+    assert bad.returncode != 0
 
 
 def test_youtu_be_shortlinks_are_not_second_channel_candidates():
